+++ models/index.js
@@ -1,7 +1,6 @@
 const Destination = require('./destination');
-//const User = require('./user');
 
 // exporting model
 module.exports = {
     Destination,
-}
+};
