+++ public/index.ejs
@@ -91,7 +91,7 @@
                 </div>
             </div>
         </div>
-        <script src="/main2.js" type="module" async defer></script> 
+        <script src="/main.js" type="module" async defer></script> 
         <script src="https://code.jquery.com/jquery-3.6.1.min.js"></script>
         <script src="https://requirejs.org/docs/release/2.3.5/minified/require.js"></script>
     </body>
