+++ server.js
@@ -1,11 +1,3 @@
-/* 
-TODO
-    REFACTOR MAIN.JS TO FIT DATABASE
-    POSSIBLY IMPLEMENT A USER DB FOR LOGGING IN AND SAVING WISHLISTS
-    IMPLEMENT OTHER FUNCTIONS (UPDATE, DELETE, ETC..)
-    RENDER EJS
-*/ 
-
 const express = require('express');
 const app = express();
 const PORT = process.env.PORT || 42373;
@@ -13,20 +5,13 @@ const db = require('./db');
 const { DestinationRoutes } = require("./routes");
 const bodyParser = require('body-parser');
 
-
+// Express request handlers //
 app.set('view engine', 'ejs');
 
-// express request handlers //
-// parse application/x-www-form-urlencoded
 app.use(bodyParser.urlencoded({ extended: true }))
-
-// parse application/json
 app.use(bodyParser.json())
-
 app.use('/api', DestinationRoutes);
 app.use(express.static("public"));
-
-app.set('view engine', 'ejs');
 
 app.get('/', (req, res) => {
      db.collection('wishlists')
